Save the Currencies into DataBase

+++ src/main/java/calculator/calculate/DepositCalculate.java
@@ -1,5 +1,5 @@
 package calculator.calculate;
 
 public interface DepositCalculate {
-    double depositProfit(double initValue, double percent, int countDay);
+    double depositProfit(double initValue, double percent, double countDay);
 }
